requirements.txt
fastapi
uvicorn
jinja2
python-multipart
passlib[bcrypt]
fpdf2
python-dotenv
sqlalchemy
psycopg2-binary
itsdangerous
alembic
